Implemented vector searching! Able to search for similar songs based on lyrics now. Auto create embeddings on insertion. Need to implement input validation.

+++ src/main/java/IEmbedder.java
@@ -1,4 +1,4 @@
 import java.util.List;
 public interface IEmbedder {
-    public double[] getEmbedding(String text);
+    public float[] getEmbedding(String text);
 }
